feat(reporting): consolidate icon components by integrating custom icons into Icons.tsx and removing insights-icons.tsx
