+++ frontend/src/features/reporting/components/RecommendationsSection.tsx
@@ -7,7 +7,7 @@ import {
   NutrientIcon,
   ProteinIcon,
   ClipboardIcon,
-} from "./insights-icons";
+} from "@/components/Icons";
 
 interface RecommendationsSectionProps {
   insights: InsightsData;
