+++ frontend/src/features/reporting/components/index.ts
@@ -1,6 +1,4 @@
-export { default as DateRangeSelector } from "@/components/chart/DateRangeSelector";
 export { default as MacroSummaryStats } from "./MacroSummaryStats";
-export { default as LineChartComponent } from "@/components/chart/LineChartComponent";
 export { default as NutritionInsights } from "./NutritionInsights";
 export { default as MealTimeBreakdown } from "./MealTimeBreakdown";
 export { default as NutrientDensityVisualization } from "./NutrientDensityVisualization";
